// src/CommandBlock.Frontend/src/app/api/model/fileEntry.ts
/**
 * CommandBlock.API | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BackupEntryDtoSizeBytes } from './backupEntryDtoSizeBytes';


export interface FileEntry { 
    name: string;
    isDirectory: boolean;
    size: BackupEntryDtoSizeBytes;
}